requirements.txt
asgiref==3.3.0
Django==3.1.3
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
importlib-metadata==2.0.0
Markdown==3.3.3
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
zipp==3.4.0
